requirements.txt
absl-py==1.4.0
accelerate==0.18.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrdict==2.0.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bed-reader==0.2.29
bgen-reader==4.0.8
bleach==6.0.0
bokeh==2.4.3
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.0
cbgen==1.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.5
coloredlogs==15.0.1
cryptography==37.0.1
cycler==0.11.0
Cython==0.29.33
cyvcf2==0.30.18
dask==2022.2.0
datasets==2.7.1
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diffusers==0.14.0
dill==0.3.6
dpcpp-cpp-rt==2022.2.1
einops==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastjsonschema==2.16.2
fastlmm==0.6.4
fastlmmclib==0.0.2
filelock==3.8.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.3
google-auth==2.16.2
google-auth-oauthlib==0.4.6
graphtools==1.5.3
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
hdbscan==0.8.29
huggingface-hub==0.15.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.3
importlib-metadata==5.1.0
importlib-resources==5.10.2
iniconfig==2.0.0
intel-cmplr-lib-rt==2022.2.1
intel-cmplr-lic-rt==2022.2.1
intel-opencl-rt==2022.2.1
intel-openmp==2022.2.1
ipdb==0.13.13
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-bokeh==3.0.5
jupyter-console==6.4.4
jupyter-server==1.23.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lightning-utilities==0.8.0
llvmlite==0.39.1
locket==1.0.0
logger==1.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mkl==2022.2.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==9.1.0
multidict==6.0.3
multiprocess==0.70.14
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.3
numpy==1.21.5
nvidia-cuda-runtime-cu11==11.7.99
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.27.8
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pandas-stubs==1.2.0.62
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pexpect==4.8.0
phate==1.0.10
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.15.0
pluggy==1.0.0
pooch==1.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyGSP==0.5.1
pynndescent==0.5.8
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
pysnptools==0.5.8
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
pytorch-lightning==1.9.4
pytz==2022.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
s-gd2==1.8.1
safetensors==0.3.1
scikit-learn==1.0.2
scipy==1.7.3
scprep==1.2.2
seaborn==0.12.2
selenium==4.11.2
Send2Trash==1.8.0
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
statsmodels==0.13.5
tasklogger==1.2.0
tbb==2021.7.1
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torchmetrics==0.11.4
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.30.2
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.5.0
umap-learn==0.5.3
urllib3==1.26.11
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
wsproto==1.2.0
xarray==0.20.2
xxhash==3.1.0
yarl==1.8.2
zipp==3.11.0
